+++ src/main/java/org/sasanlabs/fileupload/FileUploadScanRule.java
@@ -53,21 +53,29 @@ public class FileUploadScanRule extends AbstractAppParamPlugin {
 
     @Override
     public void init() {
+        if (!this.isConfigured()) {
+            getParent()
+                    .pluginSkipped(
+                            this,
+                            FileUploadI18n.getMessage(
+                                    "fileupload.configuration.not.present.skipping.scanrule"));
+            return;
+        }
         switch (this.getAttackStrength()) {
             case LOW:
-                maxRequestCount = 30;
+                maxRequestCount = 75;
                 break;
             case MEDIUM:
-                maxRequestCount = 60;
-                break;
-            case HIGH:
-                maxRequestCount = 90;
-                break;
-            case INSANE:
                 maxRequestCount = 150;
                 break;
+            case HIGH:
+                maxRequestCount = 250;
+                break;
+            case INSANE:
+                maxRequestCount = 450;
+                break;
             default:
-                maxRequestCount = 60;
+                maxRequestCount = 150;
                 break;
         }
     }
@@ -114,7 +122,7 @@ public class FileUploadScanRule extends AbstractAppParamPlugin {
                     }
                 }
             }
-            if (isMultipart && isConfigured()) {
+            if (isMultipart) {
                 FileUploadAttackExecutor fileUploadAttackExecutor =
                         new FileUploadAttackExecutor(
                                 this, nameValuePairs, originalFileName, originalContentType);